backend/lambda/requirements.txt
requests==2.23.0
SQLAlchemy==1.3.23
pg8000==1.16.2
google-cloud-secret-manager==2.4.0
faunadb==4.1.1